+++ db/upgrade.php
@@ -79,5 +79,16 @@ function xmldb_insightjournal_upgrade($oldversion) {
         upgrade_mod_savepoint(true, 2026070900, 'insightjournal');
     }
 
+    if ($oldversion < 2026070901) {
+        $table = new xmldb_table('insightjournal');
+        $field = new xmldb_field('entriesvisibility', XMLDB_TYPE_INTEGER, '1', null, XMLDB_NOTNULL, null, '0', 'completionentries');
+
+        if (!$dbman->field_exists($table, $field)) {
+            $dbman->add_field($table, $field);
+        }
+
+        upgrade_mod_savepoint(true, 2026070901, 'insightjournal');
+    }
+
     return true;
 }