+++ mod_form.php
@@ -80,6 +80,15 @@ class mod_insightjournal_mod_form extends moodleform_mod {
         $mform->setDefault('maxchars', 0);
         $mform->addHelpButton('maxchars', 'maxchars', 'insightjournal');
 
+        $mform->addElement('select', 'entriesvisibility', get_string('entriesvisibility', 'insightjournal'), [
+            INSIGHTJOURNAL_VISIBILITY_SITEDEFAULT => get_string('entriesvisibility_sitedefault', 'insightjournal'),
+            INSIGHTJOURNAL_VISIBILITY_VISIBLE => get_string('entriesvisibility_visible', 'insightjournal'),
+            INSIGHTJOURNAL_VISIBILITY_PRIVATE => get_string('entriesvisibility_private', 'insightjournal'),
+        ]);
+        $mform->setType('entriesvisibility', PARAM_INT);
+        $mform->setDefault('entriesvisibility', INSIGHTJOURNAL_VISIBILITY_SITEDEFAULT);
+        $mform->addHelpButton('entriesvisibility', 'entriesvisibility', 'insightjournal');
+
         $this->standard_coursemodule_elements();
         $this->add_action_buttons();
     }